requirements.txt
Django==3.0.2
django-background-tasks==1.2.5
django-crispy-forms==1.9.0
